refactor: input 관련 함수 분리
